somewhat final bundler config
