+++ .npmignore
@@ -1,2 +1,3 @@
 examples
 test
+build/
