+++ package.json
@@ -9,8 +9,9 @@
     "test": "mocha -r ts-node/register --require jsdom-global/register --exit test/**/*.spec.ts",
     "test-debug": "mocha -r ts-node/register --require jsdom-global/register --inspect-brk test/**/*.spec.ts",
     "typecheck": "tsc --noEmit",
-    "build": "tsc && webpack && rm -rf build",
-    "prepare": "yarn build && cp package.json README.md .gitignore .reposyncignore ./dist",
+    "tsc": "tsc",
+    "build": "tsc --emitDeclarationOnly && webpack && rm -rf ./declarations",
+    "prepare": "yarn build && cp package.json README.md .gitignore ./dist",
     "dev": "tsc-watch",
     "docs:dev": "vuepress dev docs",
     "docs:build": "vuepress build docs --dest ./docs/dist/",
